+++ CHANGELOG.md
@@ -20,9 +20,6 @@
 - Supression des jobs sur les documents si on en réinscrit un nouveau #305
 - Correction d'un problème de comptage du nombre de propriété dans le connecteur SEDA-NG #304
 - Suppression du message d'erreur et ajout de la redirection vers la page demandée lors de l'authentification CAS #363
-
-
-
   
 ## Évolutions
 
@@ -34,6 +31,10 @@
 - Ajout de la possibilité de supprimer le document Pastell une fois archivé sur le SAE pour les flux helios-generique et helios-automatique
 - La surveillance du démon prend en compte les jobs uniquement si ceux-ci sont en retard et qu'ils ont tourné au moins une fois
 - Génération SEDA des PES : si le LibelleCodBud n'est pas disponible, on mets le CodCol à la place
+- Connecteur glaneur-local : #346 ~Connecteur
+    - désactivation du traitement du glaneur en cas d'erreur lors de la suppression ou du déplacement du fichier récupéré
+    - lister le contenu des répértoires
+    - permettre le test via un fichier exemple
 
 ## Ajouts